move validator classes into their own file in viz module, update tests and utils accordingly
